ai_agent/agent.py
```python
from .nlp import process_input
from .personality import Personality
from .memory import Memory
from .code_generator import generate_code
from .internet_access import access_internet

class Agent:
    def __init__(self):
        self.user_input = ""
        self.agent_response = ""
        self.personality = Personality()
        self.memory = Memory()

    def receive_input(self, input):
        self.user_input = input

    def process_and_respond(self):
        processed_input = process_input(self.user_input)
        if processed_input.get('intent') == 'code_generation':
            self.agent_response = generate_code(processed_input.get('details'))
        else:
            self.agent_response = self.personality.generate_response(processed_input)
        self.memory.store_interaction(self.user_input, self.agent_response)

    def learn_from_internet(self):
        new_information = access_internet()
        self.memory.store_information(new_information)
        self.personality.update(new_information)

    def get_response(self):
        return self.agent_response
```