ai_agent/tests/test_agent.py
```python
import unittest
from ai_agent.agent import Agent

class TestAgent(unittest.TestCase):

    def setUp(self):
        self.agent = Agent()

    def test_process_input(self):
        user_input = "Hello, AI agent!"
        self.agent.process_input(user_input)
        self.assertEqual(self.agent.user_input, user_input)

    def test_generate_response(self):
        user_input = "Hello, AI agent!"
        self.agent.process_input(user_input)
        response = self.agent.generate_response()
        self.assertIsNotNone(response)
        self.assertIsInstance(response, str)

    def test_access_internet(self):
        url = "https://www.example.com"
        content = self.agent.access_internet(url)
        self.assertIsNotNone(content)

    def test_generate_code(self):
        code_request = "Generate a Python function to add two numbers."
        code = self.agent.generate_code(code_request)
        self.assertIsNotNone(code)
        self.assertIsInstance(code, str)

if __name__ == '__main__':
    unittest.main()
```